fix(rate-limits): address PR review feedback
- Use startsWith() for more robust endpoint categorization
- Make ban message warmer with Grove voice
- Extract test mock to shared test-utils.ts
- Document KV race conditions and DO migration path
