+++ packages/engine/src/lib/server/rate-limits/abuse.test.ts
@@ -11,27 +11,7 @@ import {
 	clearAbuseState,
 	type AbuseState
 } from './abuse.js';
-
-// Mock KV implementation
-function createMockKV(): KVNamespace {
-	const store = new Map<string, string>();
-	return {
-		get: vi.fn(async (key: string, type?: string) => {
-			const value = store.get(key);
-			if (!value) return null;
-			if (type === 'json') return JSON.parse(value);
-			return value;
-		}),
-		put: vi.fn(async (key: string, value: string) => {
-			store.set(key, value);
-		}),
-		delete: vi.fn(async (key: string) => {
-			store.delete(key);
-		}),
-		list: vi.fn(),
-		getWithMetadata: vi.fn()
-	} as unknown as KVNamespace;
-}
+import { createMockKV } from './test-utils.js';
 
 describe('getAbuseState', () => {
 	let mockKV: KVNamespace;
